Add category and sub-category update modals
Introduced UpdateCategory and UpdateSubCategory components and integrated modals for editing categories and sub-categories in the category management page. Improved error handling in user deletion logic and enhanced UI consistency for dialogs and table actions.

+++ app/users/page.tsx
@@ -100,6 +100,8 @@ export default function UsersList() {
 
   // delete user
   const handleDeleteUser = async () => {
+    if (!userToDelete) return;
+
     try {
       setDeleteLoading(true);
       const { data } = await axios.delete<UsersResponse>(
@@ -112,8 +114,15 @@ export default function UsersList() {
       } else {
         toast.error(data.message);
       }
-    } catch (error: any) {
-      toast.error(error.message);
+    } catch (error: unknown) {
+      if (axios.isAxiosError(error) && error.response) {
+        const errorMessage =
+          error.response.data.message || "Something went wrong";
+        toast.error(errorMessage);
+      } else {
+        console.error(error);
+        toast.error("Something went wrong");
+      }
     } finally {
       setDeleteLoading(false);
     }
@@ -311,9 +320,10 @@ export default function UsersList() {
         </TableBody>
       </Table>
 
-      {/* Add user modal */}
       <Dialog open={updateUserModalOpen} onOpenChange={setUpdateUserModalOpen}>
-        <DialogContent className="sm:max-w-[1100px] w-full max-h-[90vh] overflow-y-auto">
+        <DialogContent
+        onOpenAutoFocus={(e) => e.preventDefault()} 
+        className="sm:max-w-[1100px] w-full max-h-[90vh] overflow-y-auto">
           <DialogHeader>
             <span className="text-xl font-bold">Update User</span>
           </DialogHeader>
@@ -331,7 +341,6 @@ export default function UsersList() {
         </DialogContent>
       </Dialog>
 
-      {/* User Delete Alert */}
       <AlertDialog
         open={!!userToDelete}
         onOpenChange={(open) => !open && setUserToDelete(null)}
